requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.3
django-choices==1.7.0
django-crispy-forms==1.7.2
django-model-utils==3.2.0
future==0.17.1
idna==2.8
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.1
oauthlib==3.0.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pyTelegramBotAPI==3.6.6
python-telegram-bot==12.0.0b1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
